requirements.txt
psycopg2-binary==2.9.10
faker==37.1.0
pandas==2.2.3
matplotlib==3.10.1
tabulate==0.9.0
openai==1.7.0
python-dotenv==1.0.1 